added multiple computation support

+++ src/field_map.rs
@@ -1,30 +1,30 @@
-use num::Complex;
+use std::f32;
 
+use num::Complex;
 
 pub struct FieldMap {
     pub re_resolution: usize,
     pub im_resolution: usize,
-    pub precomputed_re: Vec<f64>,
-    pub precomputed_im: Vec<f64>,
+    pub precomputed_re: Vec<f32>,
+    pub precomputed_im: Vec<f32>,
 }
-
 
 impl FieldMap {
     pub fn new(
-        upper_left: Complex<f64>,
-        lower_right: Complex<f64>,
+        upper_left: Complex<f32>,
+        lower_right: Complex<f32>,
         re_resolution: usize,
         im_resolution: usize,
     ) -> Self {
-        let re_delta = (lower_right.re - upper_left.re) / re_resolution as f64;
-        let im_delta = (upper_left.im - lower_right.im) / im_resolution as f64;
+        let re_delta = (lower_right.re - upper_left.re) / re_resolution as f32;
+        let im_delta = (upper_left.im - lower_right.im) / im_resolution as f32;
 
-        let precomputed_re: Vec<f64> = (0..re_resolution)
-            .map(|x| upper_left.re + x as f64 * re_delta)
+        let precomputed_re: Vec<f32> = (0..re_resolution)
+            .map(|x| upper_left.re + x as f32 * re_delta)
             .collect();
 
-        let precomputed_im: Vec<f64> = (0..im_resolution)
-            .map(|y| upper_left.im - y as f64 * im_delta)
+        let precomputed_im: Vec<f32> = (0..im_resolution)
+            .map(|y| upper_left.im - y as f32 * im_delta)
             .collect();
 
         FieldMap {
@@ -34,9 +34,9 @@ impl FieldMap {
             precomputed_im,
         }
     }
-    
+
     #[inline(always)]
-    pub fn get_point(&self, index: usize) -> Complex<f64> {
+    pub fn get_point(&self, index: usize) -> Complex<f32> {
         let (x, y) = (index % self.re_resolution, index / self.re_resolution);
 
         Complex::new(self.precomputed_re[x], self.precomputed_im[y])
